// apps/businessbynumbers-old-vite/src/API.ts
export type ModelStringInput = {
  ne?: string | null
  eq?: string | null
  le?: string | null
  lt?: string | null
  ge?: string | null
  gt?: string | null
  contains?: string | null
  notContains?: string | null
  between?: Array<string | null> | null
  beginsWith?: string | null
  attributeExists?: boolean | null
  attributeType?: ModelAttributeTypes | null
  size?: ModelSizeInput | null
}
// ...
export type Todo = {
  __typename: 'Todo'
  id: string
  name: string
  description?: string | null
  createdAt: string
  updatedAt: string
}

export type UpdateTodoInput = {
  id: string
  name?: string | null
  description?: string | null
}

export type DeleteTodoInput = {
  id: string
}

export type ModelTodoFilterInput = {
  id?: ModelIDInput | null
  name?: ModelStringInput | null
  description?: ModelStringInput | null
  and?: Array<ModelTodoFilterInput | null> | null
  or?: Array<ModelTodoFilterInput | null> | null
  not?: ModelTodoFilterInput | null
}

export type ModelIDInput = {
  ne?: string | null
  eq?: string | null
  le?: string | null
  lt?: string | null
  ge?: string | null
  gt?: string | null
  contains?: string | null
  notContains?: string | null
  between?: Array<string | null> | null
  beginsWith?: string | null
  attributeExists?: boolean | null
  attributeType?: ModelAttributeTypes | null
  size?: ModelSizeInput | null
}

export type ModelTodoConnection = {
  __typename: 'ModelTodoConnection'
  items: Array<Todo | null>
  nextToken?: string | null
}
// ...
export type ListTodosQueryVariables = {
  filter?: ModelTodoFilterInput | null
  limit?: number | null
  nextToken?: string | null
}

export type ListTodosQuery = {
  listTodos?: {
    __typename: 'ModelTodoConnection'
    items: Array<{
      __typename: 'Todo'
      id: string
      name: string
      description?: string | null
      createdAt: string
      updatedAt: string
    } | null>
    nextToken?: string | null
  } | null
}